Wrong % of services/hosts with status OK measure

+++ shinken/misc/datamanager.py
@@ -275,7 +275,7 @@ class DataManager(object):
         if len(all_services) == 0:
             res = 0
         else:
-            res = (100-(len(problem_services) *100)/len(all_services))
+            res = int(100-(len(problem_services) *100)/float(len(all_services)))
         return res
               
     # Get percent of all Hosts
@@ -286,7 +286,7 @@ class DataManager(object):
         if len(all_hosts) == 0:
             res = 0
         else:
-            res = (100-(len(problem_hosts) *100)/len(all_hosts))
+            res = int(100-(len(problem_hosts) *100)/float(len(all_hosts)))
         return res
